♻️ Refactor: separate add day logic from useCourseStore
