+++ src/components/MyRouteContent/mockMyRoute.ts
@@ -73,7 +73,7 @@ export const mockMyRoute: MockMyRouteData = {
   ],
 };
 
-import { CourseData } from '@/src/utils/zustand/useCourseStore';
+import { CourseData } from '@/src/utils/zustand/useCourseStore/useCourseStore';
 
 export interface mockMyCourseData {
   courseName: string;
